+++ CHANGELOG.md
@@ -46,6 +46,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Header Update
 - Layout Container Update
 - Date Modified Component
+- Contributors Component
 
 ### Changed